change name of the module and some classes
