--- sample-parent/sample-features/avaiableoutput/README.adoc
@@ -1,2 +1,0 @@
-= Sample for Available Output
-
