--- sample-parent/sample-features/avaiableoutput/pom.xml
@@ -1,111 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Copyright (C) 2006-2025 Talend Inc. - www.talend.com
-   Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.talend.sdk.component</groupId>
-    <artifactId>sample-features</artifactId>
-    <version>1.80.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>avaiableoutput</artifactId>
-  <packaging>jar</packaging>
-
-  <name>Component Runtime :: Sample Feature @AvailableOutputFlows</name>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.talend.sdk.component</groupId>
-      <artifactId>component-api</artifactId>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.talend.sdk.component</groupId>
-        <artifactId>talend-component-maven-plugin</artifactId>
-        <version>${project.version}</version>
-        <!-- Please look at the TCK documentation to understand the talend-component-maven-plugin -->
-        <!-- https://talend.github.io/component-runtime/main/latest/build-tools-maven.html -->
-        <executions>
-          <execution>
-            <id>talend-dependencies</id>
-            <goals>
-              <goal>dependencies</goal>
-            </goals>
-            <phase>process-classes</phase>
-          </execution>
-          <execution>
-            <id>talend-component-bundle</id>
-            <goals>
-              <goal>car</goal>
-            </goals>
-            <phase>package</phase>
-            <configuration>
-              <excludeArtifactsFilter>com.fasterxml.jackson.core:jackson-databind,com.fasterxml.jackson.core:jackson-core,com.fasterxml.jackson.core:jackson-annotations</excludeArtifactsFilter>
-            </configuration>
-          </execution>
-          <execution>
-            <id>talend-scan-descriptor</id>
-            <goals>
-              <goal>scan-descriptor</goal>
-            </goals>
-          </execution>
-          <execution>
-            <id>talend-svg2png</id>
-            <goals>
-              <goal>svg2png</goal>
-            </goals>
-          </execution>
-          <execution>
-            <id>talend-component-validate</id>
-            <goals>
-              <goal>validate</goal>
-            </goals>
-            <phase>process-classes</phase>
-            <configuration>
-              <validatePlaceholder>true</validatePlaceholder>
-              <validateSvg>true</validateSvg>
-              <validateLegacyIcons>false</validateLegacyIcons>
-              <validateFamily>true</validateFamily>
-              <validateSerializable>true</validateSerializable>
-              <validateInternationalization>true</validateInternationalization>
-              <validateModel>true</validateModel>
-              <validateMetadata>true</validateMetadata>
-              <validateComponent>true</validateComponent>
-              <validateDataStore>true</validateDataStore>
-              <validateDataSet>true</validateDataSet>
-              <validateActions>true</validateActions>
-              <validateDocumentation>true</validateDocumentation>
-              <validateWording>true</validateWording>
-              <validateLayout>true</validateLayout>
-              <validateOptionNames>true</validateOptionNames>
-              <validateOutputConnection>true</validateOutputConnection>
-              <validateLocalConfiguration>true</validateLocalConfiguration>
-              <validateNoFinalOption>true</validateNoFinalOption>
-              <validateExceptions>true</validateExceptions>
-              <failOnValidateExceptions>true</failOnValidateExceptions>
-              <validateRecord>true</validateRecord>
-              <validateSchema>true</validateSchema>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-</project>
