+++ sample-parent/sample-features/conditional-outputs/src/main/java/org/talend/sdk/component/sample/feature/availableoutput/output/AvailableOutputNoAnnotation.java
