+++ sample-parent/sample-features/conditional-outputs/src/main/java/org/talend/sdk/component/sample/feature/availableoutput/output/ConfigurationTwo.java
@@ -1,0 +1,42 @@
+/**
+ * Copyright (C) 2006-2025 Talend Inc. - www.talend.com
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.talend.sdk.component.sample.feature.availableoutput.output;
+
+import java.io.Serializable;
+
+import org.talend.sdk.component.api.configuration.Option;
+import org.talend.sdk.component.api.configuration.ui.layout.GridLayout;
+import org.talend.sdk.component.api.meta.Documentation;
+
+import lombok.Data;
+
+@Data
+@GridLayout({
+        @GridLayout.Row({ "showSecond2" }),
+        @GridLayout.Row({ "showThird2" })
+})
+@Documentation("TODO fill the documentation for this configuration")
+public class ConfigurationTwo implements Serializable {
+
+    @Option
+    @Documentation("Show second output or not.")
+    private boolean showSecond2;
+
+    @Option
+    @Documentation("Show third output or not.")
+    private boolean showThird2;
+
+}
