+++ sample-parent/sample-features/conditional-outputs/src/main/resources/org/talend/sdk/component/sample/feature/availableoutput/Messages.properties
@@ -1,0 +1,21 @@
+#  Copyright (C) 2006-2025 Talend Inc. - www.talend.com
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+# Here you can change all your configuration display names to use more explicit labels
+# You can also translate your configuration by adding one file by local Messages_fr.properties for french for example
+sampleAvailableOutput._displayName=sampleAvailableOutput
+sampleAvailableOutput.dataset.CustomDataset._displayName=CustomDataset
+sampleAvailableOutput.datastore.CustomDatastore._displayName=CustomDatastore
+sampleAvailableOutput.actions.available_output.active_flows._displayName=Active_flows
+sampleAvailableOutput.actions.available_output.output-flow1._displayName=output flow 1
+sampleAvailableOutput.actions.available_output.output-flow2._displayName=output flow 2
