+++ sample-parent/sample-features/conditional-outputs/src/main/resources/org/talend/sdk/component/sample/feature/availableoutput/output/Messages.properties
@@ -1,0 +1,24 @@
+#  Copyright (C) 2006-2025 Talend Inc. - www.talend.com
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+# Here you can change all your configuration display names to use more explicit labels
+# You can also translate your configuration by adding one file by local Messages_fr.properties for french for example
+Configuration.dataset._displayName=Dataset
+Configuration.showSecond._displayName=Show Second
+Configuration.showThird._displayName=Show Third
+sampleAvailableOutput.Output1._displayName=Available Outputs 1.
+ConfigurationTwo.dataset._displayName=dataset
+ConfigurationTwo.showSecond2._displayName=show Second2
+ConfigurationTwo.showThird2._displayName=show Third2
+sampleAvailableOutput.Output2._displayName=Available Outputs 2.
+sampleAvailableOutput.Output0._displayName=Available Outputs without annotation
